+++ ui/components/filters/Filters.tsx
@@ -24,7 +24,7 @@ import CloneIcon from '../../public/static/img/CloneIcon';
 import ConfigurationSubscription from '@/graphql/subscriptions/ConfigurationSubscription';
 import fetchCatalogFilter from '@/graphql/queries/CatalogFilterQuery';
 import { iconMedium } from '../../css/icons.styles';
-import { RJSFModalWrapper } from '../General/Modals/Modal';
+import { RJSFModalWrapper } from '../shared/Modal/Modal';
 import { getUnit8ArrayDecodedFile, modifyRJSFSchema } from '../../utils/utils';
 import Filter from '../../public/static/img/drawer-icons/filter_svg';
 import { getMeshModels } from '../../api/meshmodel';
