+++ ui/components/filters/FiltersGrid.tsx
@@ -8,8 +8,8 @@ import {
   GridNoPapperStyles,
   GridNoTextStyles,
   GridPaginationStyles,
-} from '../designs/patterns/Grid.styles';
-import { RJSFModalWrapper } from '../General/Modals/Modal';
+} from '../MesheryPatterns/Grid.styles';
+import { RJSFModalWrapper } from '../shared/Modal/Modal';
 import CAN from '@/utils/can';
 import { keys } from '@/utils/permission_constants';
 import { Modal as SistentModal } from '@sistent/sistent';
