+++ ui/components/workspaces/index.tsx
@@ -33,11 +33,11 @@ import {
   useGetWorkspacesQuery,
   useUnassignTeamFromWorkspaceMutation,
   useUpdateWorkspaceMutation,
-} from '@/rtk-query/workspace';
-import { useNotification, useNotificationHandlers } from '@/utils/hooks/useNotification';
-import { RJSFModalWrapper } from '@/components/General/Modals/Modal';
-import _PromptComponent from '@/components/PromptComponent';
-import { EVENT_TYPES } from 'lib/event-types';
+} from '../../../rtk-query/workspace';
+import { useNotification, useNotificationHandlers } from '../../../utils/hooks/useNotification';
+import { RJSFModalWrapper } from '../../shared/Modal/Modal';
+import _PromptComponent from '../../PromptComponent';
+import { EVENT_TYPES } from '../../../lib/event-types';
 import { keys } from '@/utils/permission_constants';
 import CAN from '@/utils/can';
 import { ToolWrapper } from '@/assets/styles/general/tool.styles';
